migrate TypeScript, collapse aliases into classes (for now)
