+++ src/app/app.module.ts
@@ -75,7 +75,6 @@ import { SchoolPercentagePipe } from './pipes/school-percentage.pipe';
 import { ValuesPipe } from './pipes/values.pipe';
 import { WeatherIconPipe } from './pipes/weather-icon.pipe';
 
-import { AliasesComponent } from './components/settings/aliases/aliases.component';
 import { BackgroundComponent } from './components/settings/background/background.component';
 import { ChangePasswordComponent } from './components/settings/change-password/change-password.component';
 import { ClassesComponent } from './components/settings/classes/classes.component';
@@ -133,7 +132,6 @@ import { UrlComponent } from './components/settings/url/url.component';
 
 		// Components for settings
 		// TODO: Move these to wherever they should go
-		AliasesComponent,
 		BackgroundComponent,
 		ChangePasswordComponent,
 		ClassesComponent,
