--- src/app/components/settings/aliases/aliases.component.scss
@@ -1,19 +1,0 @@
-padding: 1rem;
-color: #55595C;
-
-h3 {
-	margin-bottom: 1rem;
-}
-
-.aliases-dismiss {
-	float: right;
-}
-
-.aliases-list {
-	display: flex;
-	justify-content: space-around;
-
-	h4 {
-		font-weight: 600;
-	}
-}
