--- src/app/components/settings/aliases/aliases.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-	selector: 'mymicds-aliases',
-	templateUrl: './aliases.component.html',
-	styleUrls: ['./aliases.component.scss']
-})
-export class AliasesComponent implements OnInit {
-
-	constructor() { }
-
-	ngOnInit() {
-	}
-
-}
