+++ src/app/components/settings/classes/classes.component.html
@@ -52,7 +52,32 @@
 		</tr>
 	</tbody>
 </table>
-<mymicds-aliases *ngIf="aliasClass" class="card"></mymicds-aliases>
+<div *ngIf="aliasClass" class="aliases card">
+	<h3>Link Classes for <strong>{{ aliasClass.name }}</strong></h3>
+	<button class="aliases-dismiss btn btn-danger" (click)="dismissAliases()">
+		<i class="fa fa-times"></i>
+	</button>
+	<div class="aliases-list">
+		<div class="aliases-portal">
+			<h4>Portal Classes</h4>
+			<div *ngFor="let class of portalClasses" class="checkbox">
+				<label>
+					<input type="checkbox" (change)="aliasChange($event, 'portal', class, aliasClass._id)" [checked]="aliasChecked('portal', class, aliasClass._id)" [disabled]="aliasDisabled('portal', class, aliasClass._id)">
+					{{ class }}
+				</label>
+			</div>
+		</div>
+		<div class="aliases-canvas">
+			<h4>Canvas Classes</h4>
+			<div *ngFor="let class of canvasClasses" class="checkbox">
+				<label>
+					<input type="checkbox" (change)="aliasChange($event, 'canvas', class, aliasClass._id)" [checked]="aliasChecked('canvas', class, aliasClass._id)" [disabled]="aliasDisabled('canvas', class, aliasClass._id)">
+					{{ class }}
+				</label>
+			</div>
+		</div>
+	</div>
+</div>
 <button class="save-classes btn btn-info" (click)="saveClasses()" [disabled]="savingClasses || (!anyClassChanged() && !anyClassAdded() && !anyClassDeleted())">
 	{{ savingClasses ? 'Saving Classes...' : 'Save Changes' }}
 </button>
