+++ src/app/components/settings/classes/classes.component.scss
@@ -1,5 +1,3 @@
-overflow-x: auto;
-
 h3 {
 	display: inline-block;
 }
@@ -52,4 +50,26 @@ table {
 
 .save-classes {
 	float: right;
+}
+
+.aliases {
+	padding: 1rem;
+	color: #55595C;
+
+	h3 {
+		margin-bottom: 1rem;
+	}
+
+	.aliases-dismiss {
+		float: right;
+	}
+
+	.aliases-list {
+		display: flex;
+		justify-content: space-around;
+
+		h4 {
+			font-weight: 600;
+		}
+	}
 }
