+++ src/app/components/settings/settings.component.scss
@@ -68,6 +68,10 @@
 		button:disabled {
 			cursor: not-allowed;
 		}
+
+		mymicds-classes {
+			overflow-x: auto;
+		}
 	}
 }
